refactor(action-popover): put sub-components in folders
