+++ src/components/action-popover/action-popover-divider/action-popover-divider.component.js
@@ -1,4 +1,4 @@
-import { MenuItemDivider as ActionPopoverDivider } from "./action-popover.style";
+import { MenuItemDivider as ActionPopoverDivider } from "../action-popover.style";
 
 ActionPopoverDivider.displayName = "ActionPopoverDivider";
 export default ActionPopoverDivider;
