+++ src/components/action-popover/action-popover-menu-button/action-popover-menu-button.d.ts
@@ -1,0 +1,13 @@
+export interface ActionPopoverMenuButtonProps {
+  children?: string;
+  buttonType?: string;
+  iconType?: string;
+  iconPosition?: string;
+  size?: string;
+}
+
+declare function ActionPopoverMenuButton(
+  props: ActionPopoverMenuButtonProps
+): JSX.Element;
+
+export default ActionPopoverMenuButton;
