+++ src/components/action-popover/action-popover-menu/action-popover-menu.component.js
@@ -1,10 +1,10 @@
 import React, { useCallback, useMemo, useContext } from "react";
 import PropTypes from "prop-types";
-import { Menu } from "./action-popover.style";
-import Events from "../../utils/helpers/events";
-import ActionPopoverItem from "./action-popover-item.component";
-import ActionPopoverDivider from "./action-popover-divider.component";
-import ActionPopoverContext from "./action-popover-context";
+import { Menu } from "../action-popover.style";
+import Events from "../../../utils/helpers/events";
+import ActionPopoverItem from "../action-popover-item/action-popover-item.component";
+import ActionPopoverDivider from "../action-popover-divider/action-popover-divider.component";
+import ActionPopoverContext from "../action-popover-context";
 
 const ActionPopoverMenu = React.forwardRef(
   (
