+++ src/components/action-popover/action-popover.stories.mdx
@@ -11,7 +11,7 @@ import {
   ActionPopoverMenu,
   ActionPopoverMenuButton,
 } from ".";
-import { MenuItem } from "./action-popover-item.component";
+import { MenuItem } from "./action-popover-item/action-popover-item.component";
 import { Table, TableRow, TableCell, TableHeader } from "../table";
 import { Accordion } from "../accordion";
 import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
