+++ src/components/action-popover/index.d.ts
@@ -1,5 +1,5 @@
 export { default as ActionPopover } from "./action-popover";
-export { default as ActionPopoverMenu } from "./action-popover-menu.component";
-export { default as ActionPopoverItem } from "./action-popover-item";
-export { default as ActionPopoverMenuButton } from "./action-popover-menu-button.component";
-export { default as ActionPopoverDivider } from "./action-popover-divider";
+export { default as ActionPopoverMenu } from "./action-popover-menu/action-popover-menu";
+export { default as ActionPopoverItem } from "./action-popover-item/action-popover-item";
+export { default as ActionPopoverMenuButton } from "./action-popover-menu-button";
+export { default as ActionPopoverDivider } from "./action-popover-divider/action-popover-divider";
