+++ src/components/action-popover/index.js
@@ -1,5 +1,5 @@
 export { default as ActionPopover } from "./action-popover.component";
-export { default as ActionPopoverMenu } from "./action-popover-menu.component";
-export { default as ActionPopoverItem } from "./action-popover-item.component";
-export { default as ActionPopoverMenuButton } from "./action-popover-menu-button.component";
-export { default as ActionPopoverDivider } from "./action-popover-divider.component";
+export { default as ActionPopoverMenu } from "./action-popover-menu/action-popover-menu.component";
+export { default as ActionPopoverItem } from "./action-popover-item/action-popover-item.component";
+export { default as ActionPopoverMenuButton } from "./action-popover-menu-button/action-popover-menu-button.component";
+export { default as ActionPopoverDivider } from "./action-popover-divider/action-popover-divider.component";
